# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/user": {
      "get": {
        "operationId": "query",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "where",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserQueryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      },
      "put": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        },
        "tags": [
          "user"
        ]
      }
    },
    "/api/session/{id}": {
      "get": {
        "operationId": "find",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "tags": [
          "session"
        ]
      },
      "put": {
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "tags": [
          "session"
        ]
      },
      "delete": {
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rev",
            "required": false,
            "in": "query",
            "schema": {}
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "session"
        ]
      }
    },
    "/api/session": {
      "get": {
        "operationId": "query",
        "parameters": [
          {
            "name": "take",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "where",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "orderBy",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "select",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "include",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionQueryResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "session"
        ]
      },
      "put": {
        "operationId": "create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSessionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          }
        },
        "tags": [
          "session"
        ]
      }
    },
...
  },
  "info": {
    "title": "Quasar Nest example",
    "description": "The cats API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Session": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "sessionId",
          "userId",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "object"
          },
          "salt": {
            "type": "object"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            }
          },
          "sessions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          }
        },
        "required": [
          "userId",
          "firstName",
          "lastName",
          "email",
          "password",
          "salt",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "Role": {
        "type": "object",
        "properties": {
          "roleId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRole"
            }
          }
        },
        "required": [
          "roleId",
          "name",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UserRole": {
        "type": "object",
        "properties": {
          "userRoleId": {
            "type": "string"
          },
          "userId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "roleId": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          },
          "isDeleted": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "deletedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "userRoleId",
          "userId",
          "roleId",
          "isDeleted",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "UserQueryResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "data"
        ]
      },
      "ConnectUserRoleDto": {
        "type": "object",
        "properties": {
          "userRoleId": {
            "type": "string"
          }
        },
        "required": [
          "userRoleId"
        ]
      },
      "CreateUserRolesRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectUserRoleDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "ConnectSessionDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          }
        },
        "required": [
          "sessionId"
        ]
      },
      "CreateUserSessionsRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSessionDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "object"
          },
          "salt": {
            "type": "object"
          },
          "roles": {
            "$ref": "#/components/schemas/CreateUserRolesRelationInputDto"
          },
          "sessions": {
            "$ref": "#/components/schemas/CreateUserSessionsRelationInputDto"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "email",
          "password",
          "salt"
        ]
      },
      "UpdateUserRolesRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectUserRoleDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateUserSessionsRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSessionDto"
            }
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "object"
          },
          "salt": {
            "type": "object"
          },
          "roles": {
            "$ref": "#/components/schemas/UpdateUserRolesRelationInputDto"
          },
          "sessions": {
            "$ref": "#/components/schemas/UpdateUserSessionsRelationInputDto"
          }
        }
      },
      "SessionQueryResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Session"
            }
          },
          "count": {
            "type": "number"
          }
        },
        "required": [
          "data"
        ]
      },
      "ConnectUserDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string"
          }
        },
        "required": [
          "userId"
        ]
      },
      "CreateSessionUserRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectUserDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "CreateSessionDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/CreateSessionUserRelationInputDto"
          }
        },
        "required": [
          "user"
        ]
      },
      "UpdateSessionUserRelationInputDto": {
        "type": "object",
        "properties": {
          "connect": {
            "$ref": "#/components/schemas/ConnectUserDto"
          }
        },
        "required": [
          "connect"
        ]
      },
      "UpdateSessionDto": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UpdateSessionUserRelationInputDto"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "access-token": []
    }
  ]
}